auth/requirements.txt

django-cors-headers==3.14.0
djangorestframework==3.14.0
PyJWT==2.3.0
sqlparse==0.4.2
mysqlparse
django_extensions
djangorestframework_simplejwt
